requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
dnspython>=2.2.0
requests>=2.28.0
python-whois>=0.8.0
openpyxl>=3.1.0
psycopg2-binary>=2.9.0
plotly>=5.18.0